requirements.txt
streamlit
numpy
opencv-python
scikit-learn
pillow
